Add save pretrained test (#318)
add save_pretrained test
